refactors iotabots
